Style the new marker window and restrict this feature only to logged in users
