+++ portfolio/src/main/webapp/index.html
@@ -323,7 +323,7 @@
             <textarea name="comment-text" class="comment-text-input" rows="4"
               placeholder="Enter your comment here..." required
               id="comment-text"></textarea>
-              <input class="submit-comment-button" type="submit">
+              <input class="submit-button submit-comment-button" type="submit">
               </form>
             </div>
           <div id="comments-section">
